backend/workers/peptide_qc/requirements.txt
# pymzml==2.7.2 # For reading Mass Spec files
# numpy # For fast array computation and peak searching

# # Infrastructure communication
# confluent-kafka
# boto3
# python-dotenv

confluent-kafka
minio
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
pymzml
numpy
biopython